Fix portability of test.c
This test file would only compile in C11 mode on a libc which implements
annex K, which is none of the ones I have access to.  Thankfully, that
was used only for strcpy_s(), which can be replaced with strncpy() with
no issue.

I also made it reflect argv[0] because my bcdec doesn’t have a .exe
suffix.